refactor: 패키지 구조 개선 및 관련 클래스 이동
- 클래스들을 적절한 패키지로 이동하여 코드 구조를 정리
