+++ src/test/java/racingcar/InputHandlerTest.java
@@ -2,6 +2,9 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.input.InputHandler;
+import racingcar.input.UserInput;
+import racingcar.output.OutputHandler;
 
 import java.io.ByteArrayInputStream;
