+++ src/test/java/racingcar/RandomValueGeneratorTest.java
@@ -2,6 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.util.RandomValueGenerator;
 
 import java.util.List;
