+++ src/test/java/racingcar/ValidatorTest.java
@@ -2,6 +2,8 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.input.UserInput;
+import racingcar.input.Validator;
 
 import static org.junit.jupiter.api.Assertions.*;
